# adapters/src/main/resources/static/openapi.yaml
openapi: 3.0.2
info:
  title: NFe Reader API
  contact:
    email: arnaldotecadm@hotmail.com
  description: >-
    API em construção.
    Esta API irá devolver informações relacionadas a Documentos Fiscais Processados
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
servers:
  - url: /api/v1
security:
  - clientCredentials: [ ]
  - bearerAuth: [ ]
paths:
  /analise-clientes/valor-total:
    get:
      tags:
        - analiseClientes
      summary: Under construction
      operationId: getAnaliseQualitativaValorTotal
      parameters:
        - name: startDate
          in: query
          description: startDate
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: endDate
          required: false
          schema:
            type: string
            format: date
      responses:
        200:
          description: Get analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnaliseQualitativaClienteDTOList'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: '#/components/responses/InternalServerError'
  /analise-clientes/qtd-pedidos:
    get:
      tags:
        - analiseClientes
      summary: Under construction
      operationId: getAnaliseQualitativaQtdPedidos
      parameters:
        - name: startDate
          in: query
          description: startDate
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: endDate
          required: false
          schema:
            type: string
            format: date
      responses:
        200:
          description: Get analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnaliseQualitativaClienteDTOList'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: '#/components/responses/InternalServerError'
  /analise-clientes/detalhes/{cpfCnpj}:
    get:
      tags:
        - analiseClientes
      summary: Under construction
      operationId: getDetalhesForCurvaABC
      parameters:
        - name: cpfCnpj
          in: path
          description: cpfCnpj
          required: true
          schema:
            type: string
      responses:
        200:
          description: Get analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstatisticaDestinatarioDTO'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: '#/components/responses/InternalServerError'
  /analise-produtos/estatistica-faturamento:
    get:
      tags:
        - analiseProdutos
      summary: Under construction
      operationId: getEstatisticaProdutoByFaturamento
      responses:
        200:
          description: Get analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProdutoCurvaABCDTOList'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: '#/components/responses/InternalServerError'
  /analise-produtos/dia-a-dia:
    get:
      tags:
        - analiseProdutos
      summary: Under construction
      operationId: getQtdProdutoDiaADia
      responses:
        200:
          description: Get analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PedidosDiaADiaDTOList'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: '#/components/responses/InternalServerError'
  /analise-produtos/estatistica-volume:
    get:
      tags:
        - analiseProdutos
      summary: Under construction
      operationId: getEstatisticaProdutoByVolume
      responses:
        200:
          description: Get analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProdutoCurvaABCDTOList'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: '#/components/responses/InternalServerError'
  /analise-produtos/estatistica-frequencia:
    get:
      tags:
        - analiseProdutos
      summary: Under construction
      operationId: getEstatisticaProdutoByFrequencia
      responses:
        200:
          description: Get analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProdutoCurvaABCDTOList'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: '#/components/responses/InternalServerError'
  /analise-produtos/curva-abc/{codigo}:
    get:
      tags:
        - analiseProdutos
      summary: Under construction
      operationId: getTotaisEstatisticaProduto
      parameters:
        - name: codigo
          in: path
          description: codigo
          required: true
          schema:
            type: string
      responses:
        200:
          description: Get analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProdutoCurvaABCDTO'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /curva-abc/faturamento-anual:
    get:
      tags:
        - curvaAbc
      summary: Under construction
      operationId: getCurvaFaturamentoAnual
      responses:
        200:
          description: Get analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicValueTotalDTOList'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: '#/components/responses/InternalServerError'
  /curva-abc/faturamento-mensal:
    get:
      tags:
        - curvaAbc
      summary: Under construction
      operationId: getCurvaFaturamentoMensal
      responses:
        200:
          description: Get analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicValueTotalDTOList'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: '#/components/responses/InternalServerError'
  /curva-abc/faturamento/estado/valor-total:
    get:
      tags:
        - curvaAbc
      summary: Under construction
      operationId: getDistVendasEstadoValorTotal
      parameters:
        - name: startDate
          in: query
          description: startDate
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: endDate
          required: false
          schema:
            type: string
            format: date
      responses:
        200:
          description: Get analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicValueTotalDTOList'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: '#/components/responses/InternalServerError'
  /curva-abc/faturamento/estado/frequencia:
    get:
      tags:
        - curvaAbc
      summary: Under construction
      operationId: getDistVendasEstadoFrequencia
      parameters:
        - name: startDate
          in: query
          description: startDate
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: endDate
          required: false
          schema:
            type: string
            format: date
      responses:
        200:
          description: Get analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicValueTotalDTOList'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: '#/components/responses/InternalServerError'
  /curva-abc/pedido/dia-a-dia:
    get:
      tags:
        - curvaAbc
      summary: Under construction
      operationId: getQtdPedidosDiaADia
      responses:
        200:
          description: Get analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PedidosDiaADiaDTOList'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /nfe/cprod/{cProd}:
    get:
      tags:
        - nfe
      summary: Under construction
      operationId: getByCodProd
      parameters:
        - name: cProd
          in: path
          description: codigo
          required: true
          schema:
            type: string
      responses:
        200:
          description: Get analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProdutoListagemNotaDTOList'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: '#/components/responses/InternalServerError'
  /nfe/cpf-cnpj/{cpfCnpj}:
    get:
      tags:
        - nfe
      summary: Under construction
      operationId: getByCpfCnpj
      parameters:
        - name: cpfCnpj
          in: path
          description: cpfCnpj
          required: true
          schema:
            type: string
      responses:
        200:
          description: Get analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProdutoListagemNotaDTOList'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      scheme: bearer
      type: http
      bearerFormat: JWT
    clientCredentials:
      type: oauth2
      description: OAuth2 Client credentials
      flows:
        clientCredentials:
          tokenUrl: ""
          scopes: { }
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiError:
      required:
        - message
      properties:
        message:
          type: string
        details:
          type: string
        validation:
          type: array
          items:
            $ref: "#/components/schemas/ValidationErrorInfo"
...
    ValidationErrorInfo:
      properties:
        field:
          type: string
        message:
          type: string
    BasicValueTotalDTOList:
      type: array
      items:
        $ref: '#/components/schemas/BasicValueTotalDTO'
    BasicValueTotalDTO:
      type: object
      properties:
        label:
          type: string
        total:
          type: number
          format: double
...
    AnaliseQualitativaClienteDTOList:
      type: array
      items:
        $ref: '#/components/schemas/AnaliseQualitativaClienteDTO'
    AnaliseQualitativaClienteDTO:
      type: object
      properties:
        cnpj:
          type: string
        razaoSocial:
          type: string
        sum:
          type: number
          format: double
        min:
          type: number
          format: double
        max:
          type: number
          format: double
        avg:
          type: number
          format: double
        count:
          type: number
          format: double
        tipoDocumento:
          type: string
    ProdutoCurvaABCDTOList:
      type: array
      items:
        $ref: '#/components/schemas/ProdutoCurvaABCDTO'
    ProdutoCurvaABCDTO:
      type: object
      properties:
        id:
          type: integer
        codigo:
          type: string
        descricao:
          type: string
        volume:
          type: number
          format: double
        faturamento:
          type: number
          format: double
        frequencia:
          type: number
          format: double
        precoMedio:
          type: number
          format: double
        ticketMedio:
          type: number
          format: double
        unidade:
          type: string
        ncm:
          type: string
        tipoDocumento:
          type: string
        partVolume:
          type: number
          format: double
        partFaturamento:
          type: number
          format: double
        partFrequencia:
          type: number
          format: double
...
    PedidosDiaADiaDTOList:
      type: array
      items:
        $ref: '#/components/schemas/PedidosDiaADiaDTO'
    PedidosDiaADiaDTO:
      type: object
      properties:
        dataEmissao:
          type: string
          format: date
        total:
          type: number
          format: double
    ProdutoListagemNotaDTOList:
      type: array
      items:
        $ref: '#/components/schemas/ProdutoListagemNotaDTO'
    ProdutoListagemNotaDTO:
      type: object
      properties:
        id:
          type: string
        naturezaOperacao:
          type: string
        numeroNotaFiscal:
          type: string
        razaoSocial:
          type: string
        chaveAcesso:
          type: string
        uf:
          type: string
        municipio:
          type: string
        valorNFe:
          type: number
          format: double
        emissao:
          type: string
          format: date
        cnpjDestinatario:
          type: string
        valorUnitario:
          type: number
          format: double
        quantidade:
          type: integer
        valorTotal:
          type: number
          format: double
        tipoDocumento:
          type: string
    EstatisticaDestinatarioDTO:
      type: object
      properties:
        razaoSocial:
          type: string
        cpfCnpj:
          type: string
        menorCompra:
          type: number
          format: double
        maiorCompra:
          type: number
          format: double
        totalPedidos:
          type: number
          format: double
        valorTotalCompras:
          type: number
          format: double
        valorMedioPedido:
          type: number
          format: double
        dataUltimaCompra:
          type: string
          format: date
        quantidadeMediaItensPorPedido:
          type: number
          format: double
        participacaoCliente:
          type: number
          format: double
        rankingFaturamento:
          type: integer
        rankingQtdPedidos:
          type: integer